+++ bundles/scripting/sightly/engine/src/main/java/org/apache/sling/scripting/sightly/impl/compiler/SightlyCompilerService.java
@@ -33,7 +33,6 @@ import org.apache.felix.scr.annotations.Service;
 import org.apache.sling.scripting.sightly.impl.compiler.debug.SanityChecker;
 import org.apache.sling.scripting.sightly.impl.compiler.ris.CommandStream;
 import org.apache.sling.scripting.sightly.impl.compiler.util.stream.PushStream;
-import org.apache.sling.scripting.sightly.impl.engine.runtime.RenderContextImpl;
 import org.apache.sling.scripting.sightly.impl.filter.Filter;
 import org.apache.sling.scripting.sightly.impl.html.dom.HtmlParserService;
 import org.apache.sling.scripting.sightly.impl.plugin.Plugin;
@@ -82,10 +81,9 @@ public class SightlyCompilerService {
      * Compile the given markup source and feed it to the given backend
      * @param source the HTML source code
      * @param backend the backend that will process the command stream from the source
-     * @param renderContext the render context
      */
-    public void compile(String source, CompilerBackend backend, RenderContextImpl renderContext) {
-        initIfNeeded(renderContext);
+    public void compile(String source, CompilerBackend backend) {
+        initIfNeeded();
         PushStream stream = new PushStream();
         SanityChecker.attachChecker(stream);
         CommandStream optimizedStream = optimizer.transform(stream);
@@ -94,13 +92,13 @@ public class SightlyCompilerService {
         frontend.compile(stream, source);
     }
 
-    private void initIfNeeded(RenderContextImpl renderContext) {
+    private void initIfNeeded() {
         if (!initialised) {
             synchronized (this) {
                 if (!initialised) {
                     ArrayList<StreamTransformer> transformers = new ArrayList<StreamTransformer>();
-                    transformers.add(ConstantFolding.transformer(renderContext));
-                    transformers.add(DeadCodeRemoval.transformer(renderContext));
+                    transformers.add(ConstantFolding.transformer());
+                    transformers.add(DeadCodeRemoval.transformer());
                     transformers.add(SyntheticMapRemoval.TRANSFORMER);
                     transformers.add(UnusedVariableRemoval.TRANSFORMER);
                     transformers.add(CoalescingWrites.TRANSFORMER);
