+++ bundles/scripting/sightly/engine/src/main/java/org/apache/sling/scripting/sightly/impl/compiler/optimization/DeadCodeRemoval.java
@@ -36,8 +36,7 @@ import org.apache.sling.scripting.sightly.impl.compiler.util.stream.Streams;
 import org.apache.sling.scripting.sightly.impl.compiler.visitor.StatefulRangeIgnore;
 import org.apache.sling.scripting.sightly.impl.compiler.visitor.StatefulVisitor;
 import org.apache.sling.scripting.sightly.impl.compiler.visitor.TrackingVisitor;
-import org.apache.sling.scripting.sightly.impl.engine.runtime.RenderContextImpl;
-import org.apache.sling.scripting.sightly.render.RenderContext;
+import org.apache.sling.scripting.sightly.impl.utils.RenderUtils;
 
 /**
  * Removes code under conditionals which are proven to fail. It is probably
@@ -47,12 +46,12 @@ import org.apache.sling.scripting.sightly.render.RenderContext;
 public class DeadCodeRemoval extends TrackingVisitor<Boolean> implements EmitterVisitor {
     // this could be merged with constant folding for better accuracy
 
-    public static StreamTransformer transformer(final RenderContext renderContext) {
+    public static StreamTransformer transformer() {
         return new StreamTransformer() {
             @Override
             public CommandStream transform(CommandStream inStream) {
                 StatefulVisitor visitor = new StatefulVisitor();
-                DeadCodeRemoval dcr = new DeadCodeRemoval(visitor.getControl(), renderContext);
+                DeadCodeRemoval dcr = new DeadCodeRemoval(visitor.getControl());
                 visitor.initializeWith(dcr);
                 Streams.connect(inStream, dcr.getOutputStream(), visitor);
                 return dcr.getOutputStream();
@@ -62,12 +61,10 @@ public class DeadCodeRemoval extends TrackingVisitor<Boolean> implements Emitter
 
     private final PushStream outStream = new PushStream();
     private final StatefulVisitor.StateControl stateControl;
-    private final RenderContextImpl renderContext;
     private final Stack<Boolean> keepConditionalEndStack = new Stack<Boolean>();
 
-    public DeadCodeRemoval(StatefulVisitor.StateControl stateControl, RenderContext renderContext) {
+    public DeadCodeRemoval(StatefulVisitor.StateControl stateControl) {
         this.stateControl = stateControl;
-        this.renderContext = (RenderContextImpl) renderContext;
     }
 
     @Override
@@ -121,16 +118,16 @@ public class DeadCodeRemoval extends TrackingVisitor<Boolean> implements Emitter
 
     private Boolean decodeConstantBool(ExpressionNode node) {
         if (node instanceof StringConstant) {
-            return renderContext.toBoolean(((StringConstant) node).getText());
+            return RenderUtils.toBoolean(((StringConstant) node).getText());
         }
         if (node instanceof BooleanConstant) {
             return ((BooleanConstant) node).getValue();
         }
         if (node instanceof NumericConstant) {
-            return renderContext.toBoolean(((NumericConstant) node).getValue());
+            return RenderUtils.toBoolean(((NumericConstant) node).getValue());
         }
         if (node instanceof NullLiteral) {
-            return renderContext.toBoolean(null);
+            return RenderUtils.toBoolean(null);
         }
         return null;
     }
