+++ bundles/scripting/sightly/engine/src/main/java/org/apache/sling/scripting/sightly/impl/engine/UnitLoader.java
@@ -33,6 +33,7 @@ import org.apache.felix.scr.annotations.Reference;
 import org.apache.felix.scr.annotations.Service;
 import org.apache.sling.api.SlingHttpServletResponse;
 import org.apache.sling.api.resource.Resource;
+import org.apache.sling.api.resource.ResourceMetadata;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.scripting.SlingBindings;
 import org.apache.sling.commons.classloader.ClassLoaderWriter;
@@ -42,12 +43,11 @@ import org.apache.sling.scripting.sightly.impl.compiled.JavaClassBackend;
 import org.apache.sling.scripting.sightly.impl.compiler.SightlyCompilerService;
 import org.apache.sling.scripting.sightly.impl.compiler.SightlyJavaCompilerService;
 import org.apache.sling.scripting.sightly.impl.compiler.SightlyParsingException;
-import org.apache.sling.scripting.sightly.impl.compiler.UnitChangeMonitor;
 import org.apache.sling.scripting.sightly.impl.compiler.util.GlobalShadowCheckBackend;
 import org.apache.sling.scripting.sightly.impl.engine.compiled.JavaClassTemplate;
 import org.apache.sling.scripting.sightly.impl.engine.compiled.SourceIdentifier;
-import org.apache.sling.scripting.sightly.impl.engine.runtime.RenderContextImpl;
 import org.apache.sling.scripting.sightly.impl.engine.runtime.RenderUnit;
+import org.apache.sling.scripting.sightly.render.RenderContext;
 import org.osgi.service.component.ComponentContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -91,26 +91,26 @@ public class UnitLoader {
      * @return the render unit
      * @throws Exception if the unit creation fails
      */
-    public RenderUnit createUnit(Resource scriptResource, Bindings bindings, RenderContextImpl renderContext) throws Exception {
+    public RenderUnit createUnit(Resource scriptResource, Bindings bindings, RenderContext renderContext) throws Exception {
+        ResourceMetadata resourceMetadata = scriptResource.getResourceMetadata();
+        String encoding = resourceMetadata.getCharacterEncoding();
+        if (encoding == null) {
+            encoding = sightlyEngineConfiguration.getEncoding();
+        }
+        SlingHttpServletResponse response = (SlingHttpServletResponse) bindings.get(SlingBindings.RESPONSE);
+        response.setCharacterEncoding(encoding);
         ResourceResolver adminResolver = renderContext.getScriptResourceResolver();
-        SourceIdentifier sourceIdentifier = new SourceIdentifier(sightlyEngineConfiguration, unitChangeMonitor, classLoaderWriter,
-            scriptResource, CLASS_NAME_PREFIX);
+        SourceIdentifier sourceIdentifier = obtainIdentifier(scriptResource);
         Object obj;
-        String encoding;
-        if (sourceIdentifier.needsUpdate()) {
-            unitChangeMonitor.touchScript(scriptResource.getPath());
-            encoding = unitChangeMonitor.getScriptEncoding(scriptResource.getPath());
-            String sourceCode = getSourceCodeForScript(adminResolver, sourceIdentifier, bindings, encoding, renderContext);
-            obj = sightlyJavaCompilerService.compileSource(sourceIdentifier, sourceCode, sourceIdentifier.getFullyQualifiedName());
+        if (needsUpdate(sourceIdentifier)) {
+            String sourceCode = getSourceCodeForScript(adminResolver, sourceIdentifier, bindings, encoding);
+            obj = sightlyJavaCompilerService.compileSource(sourceCode, sourceIdentifier.getFullyQualifiedName());
         } else {
-            encoding = unitChangeMonitor.getScriptEncoding(scriptResource.getPath());
-            obj = sightlyJavaCompilerService.getInstance(adminResolver, null, sourceIdentifier.getFullyQualifiedName(), false);
+            obj = sightlyJavaCompilerService.getInstance(adminResolver, null, sourceIdentifier.getFullyQualifiedName());
         }
         if (!(obj instanceof RenderUnit)) {
             throw new SightlyException("Class is not a RenderUnit instance");
         }
-        SlingHttpServletResponse response = (SlingHttpServletResponse) bindings.get(SlingBindings.RESPONSE);
-        response.setCharacterEncoding(encoding);
         return (RenderUnit) obj;
     }
 
@@ -121,16 +121,18 @@ public class UnitLoader {
         childTemplate = resourceFile(componentContext, CHILD_TEMPLATE_PATH);
     }
 
-    private String getSourceCodeForScript(ResourceResolver resolver, SourceIdentifier identifier, Bindings bindings, String encoding,
-                             RenderContextImpl renderContext) {
+    private SourceIdentifier obtainIdentifier(Resource resource) {
+        return new SourceIdentifier(resource, CLASS_NAME_PREFIX);
+    }
+
+    private String getSourceCodeForScript(ResourceResolver resolver, SourceIdentifier identifier, Bindings bindings, String encoding) {
         String scriptSource = null;
         try {
             Resource scriptResource = resolver.getResource(identifier.getResource().getPath());
             if (scriptResource != null) {
                 scriptSource = IOUtils.toString(scriptResource.adaptTo(InputStream.class), encoding);
-                return obtainResultSource(scriptSource, identifier, bindings, renderContext);
+                return obtainResultSource(scriptSource, identifier, bindings);
             }
-            throw new SightlyException("Cannot find template " + identifier.getResource().getPath() + " in the repository.");
         } catch (SightlyParsingException e) {
             String offendingInput = e.getOffendingInput();
             if (StringUtils.isNotEmpty(offendingInput)) {
@@ -142,22 +144,23 @@ public class UnitLoader {
                 throw e;
             }
         } catch (IOException e) {
-            throw new SightlyException("Unable to read the contents of " + identifier.getResource().getPath(), e);
+            throw new SightlyException(e);
         }
+        throw new SightlyException("Unable to generate Java class for template " + identifier.getResource().getPath());
     }
 
-    private String obtainResultSource(String scriptSource, SourceIdentifier identifier, Bindings bindings, RenderContextImpl renderContext) {
+    protected String obtainResultSource(String scriptSource, SourceIdentifier identifier, Bindings bindings) {
         JavaClassTemplate classTemplate = newMainTemplate();
         classTemplate.setClassName(identifier.getClassName());
         classTemplate.setPackageName(identifier.getPackageName());
-        CompilationOutput compilationOutput = obtainOutput(scriptSource, bindings, renderContext);
+        CompilationOutput compilationOutput = obtainOutput(scriptSource, bindings);
         processCompilationResult(compilationOutput, classTemplate);
         return classTemplate.toString();
     }
 
-    private CompilationOutput obtainOutput(String source, Bindings bindings, RenderContextImpl renderContext) {
+    private CompilationOutput obtainOutput(String source, Bindings bindings) {
         JavaClassBackend backend = new JavaClassBackend();
-        sightlyCompilerService.compile(source, new GlobalShadowCheckBackend(backend, bindings.keySet()), renderContext);
+        sightlyCompilerService.compile(source, new GlobalShadowCheckBackend(backend, bindings.keySet()));
         return backend.build();
     }
 
@@ -214,4 +217,27 @@ public class UnitLoader {
         }
         return ++line;
     }
+
+    private boolean needsUpdate(SourceIdentifier sourceIdentifier) {
+        if (sightlyEngineConfiguration.isDevMode()) {
+            return true;
+        }
+        String slyPath = sourceIdentifier.getResource().getPath();
+        String javaCompilerPath = "/" + sourceIdentifier.getFullyQualifiedName().replaceAll("\\.", "/") + ".class";
+        long javaFileDate = classLoaderWriter.getLastModified(javaCompilerPath);
+        if (javaFileDate > -1) {
+            long slyScriptChangeDate = unitChangeMonitor.getLastModifiedDateForScript(slyPath);
+            if (slyScriptChangeDate != 0) {
+                if (slyScriptChangeDate < javaFileDate) {
+                    return false;
+                }
+            } else {
+                unitChangeMonitor.touchScript(slyPath);
+            }
+            return true;
+        }
+        unitChangeMonitor.touchScript(slyPath);
+        return true;
+    }
+
 }
