+++ bundles/scripting/sightly/engine/src/main/java/org/apache/sling/scripting/sightly/impl/engine/extension/XSSRuntimeExtension.java
@@ -33,10 +33,10 @@ import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.scripting.sightly.SightlyException;
 import org.apache.sling.scripting.sightly.extension.RuntimeExtension;
 import org.apache.sling.scripting.sightly.impl.compiler.CompilerException;
-import org.apache.sling.scripting.sightly.impl.engine.runtime.RenderContextImpl;
 import org.apache.sling.scripting.sightly.impl.filter.XSSFilter;
 import org.apache.sling.scripting.sightly.impl.html.MarkupUtils;
 import org.apache.sling.scripting.sightly.impl.plugin.MarkupContext;
+import org.apache.sling.scripting.sightly.impl.utils.RenderUtils;
 import org.apache.sling.scripting.sightly.render.RenderContext;
 import org.apache.sling.xss.XSSAPI;
 import org.slf4j.Logger;
@@ -80,8 +80,7 @@ public class XSSRuntimeExtension implements RuntimeExtension {
             LOG.warn("Expression context {} is invalid, expression will be replaced by the empty string", option);
             return "";
         }
-        RenderContextImpl renderContextImpl = (RenderContextImpl) renderContext;
-        String text = renderContextImpl.toString(original);
+        String text = RenderUtils.toString(original);
         final XSSAPI xssapi = obtainAPI(renderContext.getBindings());
         return applyXSSFilter(xssapi, text, hint, markupContext);
     }
