+++ bundles/scripting/sightly/engine/src/main/java/org/apache/sling/scripting/sightly/impl/engine/extension/use/UseRuntimeExtension.java
@@ -37,8 +37,8 @@ import org.apache.felix.scr.annotations.ReferencePolicy;
 import org.apache.felix.scr.annotations.Service;
 import org.apache.sling.scripting.sightly.SightlyException;
 import org.apache.sling.scripting.sightly.extension.RuntimeExtension;
-import org.apache.sling.scripting.sightly.impl.engine.runtime.RenderContextImpl;
 import org.apache.sling.scripting.sightly.impl.plugin.UsePlugin;
+import org.apache.sling.scripting.sightly.impl.utils.RenderUtils;
 import org.apache.sling.scripting.sightly.render.RenderContext;
 import org.apache.sling.scripting.sightly.use.ProviderOutcome;
 import org.apache.sling.scripting.sightly.use.UseProvider;
@@ -72,12 +72,11 @@ public class UseRuntimeExtension implements RuntimeExtension {
         if (arguments.length != 2) {
             throw new SightlyException("Use extension requires two arguments");
         }
-        RenderContextImpl renderContextImpl = (RenderContextImpl) renderContext;
-        String identifier = renderContextImpl.toString(arguments[0]);
+        String identifier = RenderUtils.toString(arguments[0]);
         if (StringUtils.isEmpty(identifier)) {
             return null;
         }
-        Map<String, Object> useArgumentsMap = renderContextImpl.toMap(arguments[1]);
+        Map<String, Object> useArgumentsMap = RenderUtils.toMap(arguments[1]);
         Bindings useArguments = new SimpleBindings(Collections.unmodifiableMap(useArgumentsMap));
         ArrayList<UseProvider> providers = new ArrayList<UseProvider>(providersMap.values());
         ListIterator<UseProvider> iterator = providers.listIterator(providers.size());
