+++ bundles/scripting/sightly/engine/src/test/java/org/apache/sling/scripting/sightly/impl/utils/RenderUtilsTest.java
@@ -1,0 +1,7 @@
+package org.apache.sling.scripting.sightly.impl.utils;
+
+/**
+ * Created by bailescu on 9/3/15.
+ */
+public class RenderUtilsTest {
+}
